requirements.txt
langchain
langchain-community
langchain-huggingface
chromadb
sentence-transformers
pypdf
streamlit
python-dotenv
langchain_chroma
